closerandcloser

+++ tasks/mpi/strakhov_a_char_freq_counter/src/ops_mpi.cpp
@@ -61,12 +61,12 @@ bool strakhov_a_char_freq_counter_mpi::CharFreqCounterPar::RunImpl() {
   unsigned int world_size = world_.size();
   unsigned int segment = input_length / world_size;
   auto excess = input_length % world_size;
-  std::vector<int> send_counts(world_size, segment);
+  std::vector<int> send_counts(world_size, static_cast<signed int>(segment));
   for (unsigned int i = 0; i < excess; i++) {
     send_counts[i]++;
   }
   std::vector<int> displacements(world_size, 0);
-  for (int i = 1; i < world_size; i++) {
+  for (unsigned int i = 1; i < world_size; i++) {
     displacements[i] = displacements[(i - 1)] + send_counts[(i - 1)];
   }
   local_input_ = std::vector<signed char>(send_counts[rank]);
